+++ ado_wrapper/resources/artifact.py
@@ -1,3 +1,4 @@
+import json
 from dataclasses import dataclass, field
 from typing import TYPE_CHECKING, Any
 
@@ -16,10 +17,10 @@ class BuildArtifact(StateManagedResource):
 
     artifact_id: str = field(metadata={"is_id_field": True})
     name: str
-    source_job_id: str
-    artifact_local_path: str | None
-    artifact_size_bytes: int | None
-    download_path: str
+    source_job_id: str = field(repr=False)
+    artifact_local_path: str | None = field(repr=False)
+    artifact_size_bytes: int | None = field(repr=False)
+    download_path: str = field(repr=False)
     build_id: str
 
     @classmethod
@@ -52,6 +53,23 @@ class BuildArtifact(StateManagedResource):
         return f"https://dev.azure.com/{ado_client.ado_org_name}/{ado_client.ado_project_name}/_build/results?buildId={self.build_id}&view=artifacts&pathAsName=false&type=publishedArtifacts"  # fmt: skip
 
     # ===================================================================================================
+
+    @classmethod
+    def get_all_consume_by_build(cls, ado_client: "AdoClient", build_id: str) -> list["BuildArtifact"]:
+        request = ado_client.session.get(
+                f"https://dev.azure.com/{ado_client.ado_org_name}/{ado_client.ado_project_name}/_build/results?buildId={build_id}&view=artifacts&pathAsName=false&type=consumedArtifacts"
+        ).text
+        json_raw = request.split('<script id="dataProviders" type="application/json">')[1].split("</script>")[0]
+        json_data = json.loads(json_raw)
+        consumed_artifacts_data = json_data["data"]["ms.vss-build-web.run-consumed-artifacts-data-provider"]["consumedSources"]
+        pipeline_artifacts_data = [x for x in consumed_artifacts_data if x["artifactType"] == "Pipeline"]  # versionId is the build_id
+        build_id_to_artifact_names = {data["versionId"]: [x["artifactName"] for x in  data["consumedArtifacts"]] for data in pipeline_artifacts_data}
+        artifacts = []
+        for build_id, artifact_names in build_id_to_artifact_names.items():
+            build_artifacts = [Artifact.get_by_name(ado_client, build_id, artifact_name) for artifact_name in artifact_names]
+            artifacts.extend(build_artifacts)
+        return artifacts
+
 
     # @classmethod
     # def create(cls, ado_client: "AdoClient", build_id: str, artifact_name: str) -> "BuildArtifact":